refactor: 검색 결과 페이지 상품 개수 표시
