+++ src/components/common/Card/CardList/Search/index.tsx
@@ -1,3 +1,4 @@
+import { useRouter } from 'next/router';
 import { useInfiniteQuery } from '@tanstack/react-query';
 import classNames from 'classnames/bind';
 
@@ -7,6 +8,7 @@ import { infiniteProductsSearchQueries } from '@/apis/product/queries';
 import useIntersect from '@/hooks/useIntersect';
 import CardPlaceholder from '@/components/common/Card/CardPlaceholder';
 import SearchNotFound from '@/components/search/SearchNotFound';
+import SortButton from '@/components/common/Button/Sort';
 
 interface CardListSearchProps {
   className?: string;
@@ -16,9 +18,18 @@ interface CardListSearchProps {
 
 const cx = classNames.bind(styles);
 
+const SORT_OPTIONS = [
+  { name: '최신순', value: '0' },
+  { name: '별점 높은 순', value: '1' },
+  { name: '별점 낮은 순', value: '2' },
+  { name: '가격 높은 순', value: '3' },
+  { name: '가격 낮은 순', value: '4' },
+];
+
 const PAGE_SIZE = 8;
 
 export default function CardListSearch({ className, orderBy = '0', keyword }: CardListSearchProps) {
+  const router = useRouter();
   const {
     data: productsData,
     hasNextPage,
@@ -32,6 +43,7 @@ export default function CardListSearch({ className, orderBy = '0', keyword }: Ca
   const productsPages = productsData?.pages ?? [];
   const hasTargetRef = !isFetchingNextPage && hasNextPage;
   const isSearchNotFound = productsPages.length === 0 || productsPages[0].totalCount === 0;
+  const productsCount = productsPages[0]?.totalCount || 0;
 
   const targetRef = useIntersect(async (entry, observer) => {
     observer.unobserve(entry.target);
@@ -46,6 +58,23 @@ export default function CardListSearch({ className, orderBy = '0', keyword }: Ca
 
   return (
     <div className={cx('container', className)}>
+      <div className={styles.sortBox}>
+        <SortButton
+          options={SORT_OPTIONS}
+          initialOptionValue={orderBy}
+          onClick={value => {
+            router.replace({
+              pathname: '/search/result',
+              query: {
+                ...router.query,
+                orderBy: value,
+              },
+            });
+          }}
+        />
+        <p className={styles.productsCount}>{productsCount}개 상품</p>
+      </div>
+      <div className={styles.divider} />
       {!isSearchNotFound ? (
         <ul className={styles.list}>
           {productsPages.map(productsPage =>
