+++ src/pages/search/result/SearchResultPage.module.scss
@@ -22,22 +22,4 @@
   .searchBox {
     width: 100%;
   }
-
-  .sortBox {
-    display: flex;
-    width: 100%;
-    padding: 0.8rem 1.6rem;
-    justify-content: flex-end;
-  }
-
-  .divider {
-    background-color: $color-gray-50;
-    width: 100%;
-    height: 0.8rem;
-  }
-
-  .contents {
-    min-height: calc(100vh - 26.6rem);
-    padding: 2rem 0;
-  }
 }
