+++ src/pages/search/result/index.tsx
@@ -7,10 +7,8 @@ import { yupResolver } from '@hookform/resolvers/yup';
 import styles from './SearchResultPage.module.scss';
 import Header from '@/components/common/Layout/Header';
 import CartButton from '@/components/common/Button/Cart';
-import GitHubBox from '@/components/common/GitHubBox';
 import FloatingBox from '@/components/common/Layout/Footer/FloatingBox';
 import NavBottom from '@/components/common/Nav/Bottom';
-import SortButton from '@/components/common/Button/Sort';
 import useToast from '@/hooks/useToast';
 import ScrollTopButton from '@/components/common/Button/ScrollTop';
 import SearchInput from '@/components/common/Input/SearchInput';
@@ -19,14 +17,6 @@ import { SearchFormValues, searchSchema } from '@/utils/searchSchema';
 import BackButton from '@/components/common/Button/BackButton';
 import CardListSearch from '@/components/common/Card/CardList/Search';
 import FloatingActionBox from '@/components/common/Layout/Footer/FloatingActionBox';
-
-const SORT_OPTIONS = [
-  { name: '최신순', value: '0' },
-  { name: '별점 높은 순', value: '1' },
-  { name: '별점 낮은 순', value: '2' },
-  { name: '가격 높은 순', value: '3' },
-  { name: '가격 낮은 순', value: '4' },
-];
 
 const BOTTOM_BOX_ID = 'bottomBox';
 
@@ -118,26 +108,7 @@ export default function SearchResultPage({ keyword, orderBy }: ProductCategoryPa
           </Header.Right>
         </Header.Box>
       </Header.Root>
-      <div className={styles.sortBox}>
-        <SortButton
-          options={SORT_OPTIONS}
-          initialOptionValue={orderBy}
-          onClick={value => {
-            router.replace({
-              pathname: '/search/result',
-              query: {
-                ...router.query,
-                orderBy: value,
-              },
-            });
-          }}
-        />
-      </div>
-      <div className={styles.divider} />
-      <div className={styles.contents}>
-        <CardListSearch keyword={keyword} orderBy={orderBy} />
-      </div>
-      <GitHubBox />
+      <CardListSearch keyword={keyword} orderBy={orderBy} />
       <FloatingBox id={BOTTOM_BOX_ID}>
         <NavBottom />
         <FloatingActionBox>
